Adresses Clément's comments (1st part)
