+++ src/gui/parameterswidget.cpp
@@ -17,12 +17,11 @@
  */
 
 #include "parameterswidget.h"
-
 #include "customdialog.h"
 
 namespace KDC {
 
-ParametersWidget::ParametersWidget(QWidget *parent) : GuiUtility::LargeWidgetWithCustomToolTip(parent) {}
+ParametersWidget::ParametersWidget(QWidget *parent) : LargeWidgetWithCustomToolTip(parent) {}
 
 void ParametersWidget::setEnabled(bool val) {
     // Disable all parent CustomDialog
