+++ src/gui/widgetwithcustomtooltip.h
@@ -1,0 +1,59 @@
+/*
+ * Infomaniak kDrive - Desktop
+ * Copyright (C) 2023-2024 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+
+#pragma once
+
+#include <QPoint>
+#include <QString>
+#include <QWidget>
+
+class QHelpEvent;
+class QLayout;
+
+namespace KDC {
+
+class CustomToolTip;
+
+class WidgetWithCustomToolTip : public QWidget {
+        Q_OBJECT
+
+    public:
+        explicit WidgetWithCustomToolTip(QWidget *parent);
+        void setCustomToolTipText(const QString &text) { _customToolTipText = text; }
+
+    private:
+        virtual QPoint customToolTipPosition(QHelpEvent *event);
+
+    protected:
+        CustomToolTip *_customToolTip;
+        QString _customToolTipText;
+        virtual bool event(QEvent *event) override;
+        virtual void leaveEvent(QEvent *event) override;
+};
+
+class LargeWidgetWithCustomToolTip : public WidgetWithCustomToolTip {
+        Q_OBJECT
+
+    public:
+        explicit LargeWidgetWithCustomToolTip(QWidget *parent);
+
+    private:
+        virtual QPoint customToolTipPosition(QHelpEvent *event) override;
+};
+}  // namespace KDC
