+++ src/libcommon/comm.h
@@ -42,6 +42,8 @@
 
 typedef enum { REQUEST = 0, REPLY, SIGNAL } MsgType;
 
+static const QColor driveDefaultColor(0x9F9F9F);
+
 typedef enum {
     REQUEST_NUM_LOGIN_REQUESTTOKEN = 1,
     REQUEST_NUM_USER_DBIDLIST,
